Implement CircularArea And extract SimpleBoxArea and PolygonArea to higher level
